Update README.md - Adjusting to Formatting for consistency code samples
The formatting of the code samples in the README for were inconsistent. Some were really spread out and difficult to read. Using the JSON Stored Proc code sample as a guide, I updated the code to easily show hierarchy in HTML samples, JS, SQL, etc.

Point of these changes was to increase readability of code samples.